requirements.txt
sqlalchemy
pydantic
mysql-connector-python  # o pymysql si prefieres
python-jose[cryptography] passlib[bcrypt]
